a sensible source structure
The idea is that src/kernel/ is only allowed to interface with the
hardware using whatever's defined in src/arch/generic.h. I'll probably
write a small script for checking this later on.

This is a giant commit so I've probably fucked something up. It boots
fine on Bochs and QEMU, so at least there's that.