store the processes as a tree instead of a list
I'm about to need that for waits(). There's no single list of processes
for simplicity's sake, but the search will now be even slower and it
might even introduce a potential vuln! How fun! Someone could override
stuff in .bss with random values. I'll either make gcc check if it
hasn't gone past the end of the stack, or turn this into a non recursive
function.