init/fs: handle delegated reads in a subprocess
first off: this is horrible. there's no CoW, so this is horribly slow.

the sole purpose of this is to let the userland delegate handle multiple
requests at once, because i'm implementing concurrent IO and I kinda
need that.

it's possible that handling of multiple requests could be handled in a
better way - but this could be good enough once i implement CoW