shared: rename ufs_request to better fit its role in userland
The old name could have suggested that it held a response to a request
received by fs_wait. The new name is unfortunately very similar to
the `struct vfs_request` already used internally in the kernel, but
it's better at conveying that it contains a filesystem request yet to
be handled.

vfs_request - virtual filesystem request (a bad name in hindsight)
ufs_request - user filesystem request